% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linR.R
\name{linR}
\alias{linR}
\title{Linear Regression}
\usage{
linR(formula, data, level = 0.95)
}
\arguments{
\item{formula}{A formula of interest}

\item{data}{An input data}

\item{level}{Confidence level for the confidence intervals (default = .95)}
}
\value{
A list containing regression results:
\itemize{
\item \code{coefficients}: Estimated regression coefficients.
\item \code{std_error}: Standard errors of the coefficients.
\item \code{t_values}: t-statistics of the coefficients.
\item \code{p_values}: p-values for the t-tests.
\item \code{R_squared}: R-squared value.
\item \code{adj_R_squared}: Adjusted R-squared value.
\item \code{F_statistic}: F-statistic for the overall model.
\item \code{F_p_value}: p-value for the F-statistic.
\item \code{fitted_values}: Predicted values.
\item \code{residuals}: Residuals.
\item \code{degrees_of_freedom}: Degrees of freedom for the model.
\item \code{CI_table}: Confidence intervals for the coefficients.
}
}
\description{
This function performs linear regression.
}
\examples{
data(airquality)
linR(Ozone ~ Temp + Wind, data = airquality, level = 0.95)

}
